fixed few warnings
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@7410 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ motion_planning/sbpl/src/discrete_space_information/nav3dkin/environment_nav3Dkin.cpp
@@ -1097,7 +1097,7 @@ int EnvironmentNAV3DKIN::GetStartHeuristic(int stateID)
 
 	EnvNAV3DKINHashEntry_t* HashEntry = EnvNAV3DKIN.StateID2CoordTable[stateID];
 	int h2D = grid2Dsearch->getlowerboundoncostfromstart_inmm(HashEntry->X, HashEntry->Y);
-	int hEuclid = NAV3DKIN_COSTMULT_MTOMM*EuclideanDistance_m(EnvNAV3DKINCfg.StartX_c, EnvNAV3DKINCfg.StartY_c, HashEntry->X, HashEntry->Y);
+	int hEuclid = (int)(NAV3DKIN_COSTMULT_MTOMM*EuclideanDistance_m(EnvNAV3DKINCfg.StartX_c, EnvNAV3DKINCfg.StartY_c, HashEntry->X, HashEntry->Y));
 		
 
 #if DEBUG
